move some pol correction helpers to the kernel
