+++ Framework/Algorithms/src/PolarizationCorrections/PolarizationCorrectionsHelpers.cpp
@@ -7,6 +7,7 @@
 
 #include "MantidAlgorithms/PolarizationCorrections/PolarizationCorrectionsHelpers.h"
 #include "MantidAPI/Run.h"
+#include "MantidKernel/SpinStateHelpers.h"
 #include "MantidKernel/StringTokenizer.h"
 
 #include <algorithm>
@@ -28,39 +29,13 @@ namespace PolarizationCorrectionsHelpers {
  */
 API::MatrixWorkspace_sptr workspaceForSpinState(API::WorkspaceGroup_sptr group, const std::string &spinStateOrder,
                                                 const std::string &targetSpinState) {
-  const auto &spinStateOrderVec = splitSpinStateString(spinStateOrder);
-  const auto &wsIndex = indexOfWorkspaceForSpinState(spinStateOrderVec, targetSpinState);
+  const auto &spinStateOrderVec = Mantid::Kernel::SpinStateHelpers::splitSpinStateString(spinStateOrder);
+  const auto &wsIndex =
+      Mantid::Kernel::SpinStateHelpers::indexOfWorkspaceForSpinState(spinStateOrderVec, targetSpinState);
   if (!wsIndex.has_value()) {
     return nullptr;
   }
   return std::dynamic_pointer_cast<API::MatrixWorkspace>(group->getItem(wsIndex.value()));
-}
-
-/*
-For a given workspace group, spin state order, and desired spin state, this method will
-return the index of the specified workspace in the group, using the position of the desired spin
-state in the spin state order.
-*/
-std::optional<size_t> indexOfWorkspaceForSpinState(const std::vector<std::string> &spinStateOrder,
-                                                   std::string targetSpinState) {
-  boost::trim(targetSpinState);
-  size_t const index =
-      std::find(spinStateOrder.cbegin(), spinStateOrder.cend(), targetSpinState) - spinStateOrder.cbegin();
-  if (index == spinStateOrder.size()) {
-    return std::nullopt;
-  }
-  return index;
-}
-
-/*
-For a given spin state input string of the form e.g. "01,11,00,10", split the string
-into a vector of individual spin states. This will also trim any leading/trailing
-whitespace in the individual spin states.
-*/
-std::vector<std::string> splitSpinStateString(const std::string &spinStates) {
-  using Mantid::Kernel::StringTokenizer;
-  StringTokenizer tokens{spinStates, ",", StringTokenizer::TOK_TRIM};
-  return std::vector<std::string>{tokens.begin(), tokens.end()};
 }
 } // namespace PolarizationCorrectionsHelpers
