+++ Framework/Algorithms/test/PolarizationCorrections/PolarizationCorrectionsHelpersTest.h
@@ -85,34 +85,6 @@ public:
     TS_ASSERT_THROWS(SpinStatesORSO::addORSOLogForSpinState(ws, "invalidSpinState"), std::invalid_argument &);
   }
 
-  void testSplitSpinStateStringSuccessful() { runTestSplitSpinStateString("01,11,10,00", {"01", "11", "10", "00"}); }
-
-  void testSplitSpinStateStringWithSpaces() {
-    runTestSplitSpinStateString(" 01 ,  11 , 10 ,  00 ", {"01", "11", "10", "00"});
-  }
-
-  void testSplitSpinStateStringEmptyString() { runTestSplitSpinStateString("", {}); }
-
-  void testSplitSpinStateStringSingleItem() { runTestSplitSpinStateString("01", {"01"}); }
-
-  void testIndexOfWorkspaceForSpinState_TargetStateExists() {
-    runTestIndexOfWorkspaceForSpinState({"00", "11", "10", "01"}, "10", 2);
-  }
-
-  void testIndexOfWorkspaceForSpinState_TargetStateDoesNotExist() {
-    runTestIndexOfWorkspaceForSpinState({"00", "11", "10", "01"}, "invalid_state");
-  }
-
-  void testIndexOfWorkspaceForSpinState_EmptySpinStateOrder() { runTestIndexOfWorkspaceForSpinState({}, "10"); }
-
-  void testIndexOfWorkspaceForSpinState_DuplicateEntries() {
-    runTestIndexOfWorkspaceForSpinState({"10", "10", "11"}, "10", 0);
-  }
-
-  void testIndexOfWorkspaceForSpinState_TrimWhitespace() {
-    runTestIndexOfWorkspaceForSpinState({"00", "11", "10", "01"}, " 10 ", 2);
-  }
-
   void testErrorPropagation_linear() {
     constexpr size_t vars = 1;
     using Types = Arithmetic::ErrorTypeHelper<vars>;
@@ -270,22 +242,6 @@ private:
       TS_ASSERT(run.hasProperty(SpinStatesORSO::LOG_NAME));
       TS_ASSERT_EQUALS(run.getPropertyValueAsType<std::string>(SpinStatesORSO::LOG_NAME), ORSO_SPIN_STATES[i]);
     }
-  }
-
-  void runTestSplitSpinStateString(const std::string &spinStates, const std::vector<std::string> &expectedResult) {
-    std::vector<std::string> result = PolarizationCorrectionsHelpers::splitSpinStateString(spinStates);
-    TS_ASSERT_EQUALS(result.size(), expectedResult.size());
-
-    for (size_t i = 0; i < result.size(); ++i) {
-      TS_ASSERT_EQUALS(result[i], expectedResult[i]);
-    }
-  }
-
-  void runTestIndexOfWorkspaceForSpinState(const std::vector<std::string> &spinStateOrder,
-                                           const std::string &targetSpinState,
-                                           const std::optional<size_t> expectedIndex = std::nullopt) {
-    auto index = PolarizationCorrectionsHelpers::indexOfWorkspaceForSpinState(spinStateOrder, targetSpinState);
-    TS_ASSERT_EQUALS(index, expectedIndex);
   }
 
   void runTestWorkspaceForSpinState(const std::vector<std::string> &spinStateOrder, const std::string &targetSpinState,
