+++ Framework/DataHandling/src/SaveNXcanSASHelper.cpp
@@ -18,6 +18,7 @@
 #include "MantidGeometry/MDGeometry/IMDDimension.h"
 #include "MantidKernel/MDUnit.h"
 #include "MantidKernel/MantidVersion.h"
+#include "MantidKernel/SpinStateHelpers.h"
 #include "MantidKernel/StringTokenizer.h"
 #include "MantidKernel/VectorHelper.h"
 #include "MantidNexus/H5Util.h"
@@ -26,6 +27,7 @@
 #include <regex>
 
 using namespace Mantid::Kernel;
+using namespace Mantid::Kernel::SpinStateHelpers;
 using namespace Mantid::Geometry;
 using namespace Mantid::API;
 using namespace Mantid::DataHandling::NXcanSAS;
@@ -225,7 +227,7 @@ void writePolarizedDataToFile(H5::DataSet &dataSet, WorkspaceExtractorFunctor fu
   for (size_t i = 0; i < spin.pIn.size(); i++) {
     for (size_t j = 0; j < spin.pOut.size(); j++) {
       auto state = stateConverter(spin.pIn.at(i)) + stateConverter(spin.pOut.at(j));
-      auto index = findWorkspaceIndexForSpinState(spin.spinVec, state);
+      auto index = indexOfWorkspaceForSpinState(spin.spinVec, state);
       if (!index.has_value()) {
         throw std::runtime_error("Couldn't find workspace for spin state " + state);
       }
